Clarify that OAuth is but one target auth realm
This commit includes a quick pass I made over the document to serve as
an example on how we could be more general about the pattern,
acknowledging that OAuth may or may not be involved in these flows.
This more closely matches real world usage in cloud services like AWS
and GCP.

I didn't put a ton of thought into the naming, and I also noticed lots
of other spots that felt funny along the way, but refrained from further
changes to keep the scope of this sample tight. It will have some
knock-on effects as other parts of the text reference e.g. RFC 7523. I
stopped short of suggesting changes in the security considerations
section for that reason. Happy to continue work in this direction if
useful.